+++ buildSrc/src/main/kotlin/Versions.kt
@@ -1,6 +1,6 @@
 object Versions {
     const val protobuf = "3.6.1"
     const val grpc = "1.15.1"
-    const val kotlin = "1.3.0"
-    const val coroutines = "1.0.0"
+    const val kotlin = "1.3.11"
+    const val coroutines = "1.1.0"
 }